refactor: split filters into individual files
