+++ filters/default.js
@@ -1,0 +1,5 @@
+module.exports = function _default(input, def) {
+    return input !== undefined && (input || typeof input === "number")
+        ? input
+        : def;
+};
