+++ filters/json.js
@@ -1,0 +1,3 @@
+module.exports = function json(input, indent) {
+    return JSON.stringify(input, null, indent || 0);
+};
