+++ filters/lower.js
@@ -1,0 +1,10 @@
+const iterateFilter = require("./iterateFilter");
+
+module.exports = function lower(input) {
+    var out = iterateFilter.apply(lower, arguments);
+    if (out !== undefined) {
+        return out;
+    }
+
+    return input.toString().toLowerCase();
+};
