+++ filters/raw.js
@@ -1,0 +1,7 @@
+const safe = require("./safe");
+
+function raw(input) {
+    return safe(input);
+}
+raw.safe = true;
+module.exports = raw;
