+++ filters/replace.js
@@ -1,0 +1,4 @@
+module.exports = function replace(input, search, replacement, flags) {
+    var r = new RegExp(search, flags);
+    return input.replace(r, replacement);
+};
