+++ filters/reverse.js
@@ -1,0 +1,5 @@
+const sort = require("./sort");
+
+module.exports = function reverse(input) {
+    return sort(input, true);
+};
